introducing spritely
nulled single sprite example, replaced with spritely trial. so far,
works best specified in-line in head, but that may be the recurring
issue with stylesheets not asserting properties over elements with id..
fix with a class to clear header space? for now, this should still
operate faster than old code if I can halt its autoplay.